+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Version history
 
 ## 0.1.0
-- Verify required `srb`, `bundle` during extension startup.
+- Verify required `srb`, `bundle` during extension startup (can be disabled via config).
 - Initial version, forked from official 0.3.41.